+++ Makefile
@@ -69,11 +69,13 @@ notcompiler: FORCE
 	@$(MAKE) runtime
 	@$(MAKE) modules
 
-dyno: FORCE
-	@echo "Making the compiler library..."
+frontend: FORCE
+	@echo "Making the frontend compiler library..."
 	@cd third-party && $(MAKE) llvm
 	@cd third-party && $(MAKE) CHPL_MAKE_HOST_TARGET=--host jemalloc
-	@cd compiler/dyno && $(MAKE) -f Makefile.help dyno
+	@cd frontend && $(MAKE) -f Makefile.help frontend
+
+dyno: frontend FORCE
 
 compiler: FORCE
 	@echo "Making the compiler..."
@@ -83,7 +85,7 @@ compiler: FORCE
 
 parser: FORCE
 	@echo "Making the parser..."
-	@cd compiler/dyno && $(MAKE) -f Makefile.help dyno-parser
+	@cd frontend && $(MAKE) -f Makefile.help parser
 
 modules: FORCE
 	@echo "Making the modules..."
@@ -168,7 +170,7 @@ clean: FORCE
 	cd modules && $(MAKE) clean
 	cd runtime && $(MAKE) clean
 	cd third-party && $(MAKE) clean
-	cd compiler/dyno/tools/chpldoc && $(MAKE) clean
+	cd tools/chpldoc && $(MAKE) clean
 	if [ -e doc/Makefile ]; then cd doc && $(MAKE) clean; fi
 	rm -f util/chplenv/*.pyc
 
@@ -177,7 +179,7 @@ cleanall: FORCE
 	cd modules && $(MAKE) cleanall
 	cd runtime && $(MAKE) cleanall
 	cd third-party && $(MAKE) cleanall
-	cd compiler/dyno/tools/chpldoc && $(MAKE) cleanall
+	cd tools/chpldoc && $(MAKE) cleanall
 	if [ -e doc/Makefile ]; then cd doc && $(MAKE) cleanall; fi
 	rm -f util/chplenv/*.pyc
 	rm -rf build
@@ -195,7 +197,7 @@ clobber: FORCE
 	cd tools/c2chapel && $(MAKE) clobber
 	cd tools/mason && $(MAKE) clobber
 	cd tools/protoc-gen-chpl && $(MAKE) clobber
-	cd compiler/dyno/tools/chpldoc && $(MAKE) clobber
+	cd tools/chpldoc && $(MAKE) clobber
 	if [ -e doc/Makefile ]; then cd doc && $(MAKE) clobber; fi
 	rm -rf bin
 	rm -rf lib
