+++ compiler/AST/AggregateType.cpp
@@ -38,7 +38,7 @@
 #include "symbol.h"
 #include "visibleFunctions.h"
 #include "wellknown.h"
-#include "../dyno/lib/immediates/prim_data.h"
+#include "../../frontend/lib/immediates/prim_data.h"
 
 #include "global-ast-vecs.h"
