+++ compiler/include/flags.h
@@ -25,7 +25,7 @@
 //
 // HOW TO ADD A FLAG:
 // - add an entry (sorted alphabetically) to the list in
-//   compiler/dyno/include/chpl/uast/PragmaList.h
+//   frontend/include/chpl/uast/PragmaList.h
 // - to add the flag to a symbol, use Symbol::addFlag(Flag)
 // - to check whether a symbol has a flag, use Symbol::hasFlag(Flag)
 //
