+++ compiler/llvm/clangUtil.cpp
@@ -107,7 +107,7 @@
 #include "llvmDebug.h"
 #include "llvmVer.h"
 
-#include "../dyno/lib/immediates/prim_data.h"
+#include "../../frontend/lib/immediates/prim_data.h"
 
 #include "global-ast-vecs.h"
