+++ compiler/make/Makefile.compiler.head
@@ -67,7 +67,7 @@ INCL_CFLAGS = -I. \
               -I$(COMPILER_BUILD) \
               -I$(RUNTIME_ROOT)/include/encoding \
               -I$(THIRD_PARTY_DIR)/utf8-decoder \
-              -I$(COMPILER_ROOT)/dyno/include \
+              -I$(COMPILER_ROOT)/../frontend/include \
 
 
 COMP_CXXFLAGS += $(INCL_CFLAGS)
