+++ frontend/lib/immediates/Makefile.include
@@ -16,16 +16,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-DYNO_IMMEDIATES_OBJDIR = $(COMPILER_BUILD)/dyno/lib/immediates
+FRONTEND_IMMEDIATES_OBJDIR = $(COMPILER_BUILD)/frontend/lib/immediates
 
-ALL_SRCS += dyno/lib/immediates/*.h dyno/lib/immediates/*.cpp
+ALL_SRCS += ../frontend/lib/immediates/*.h ../frontend/lib/immediates/*.cpp
 
-DYNO_IMMEDIATES_SRCS =                              \
+FRONTEND_IMMEDIATES_SRCS =                          \
            ifa_vars.cpp                             \
            hash_multipliers.cpp                     \
            num.cpp                                  \
 
-SRCS = $(DYNO_IMMEDIATES_SRCS)
+SRCS = $(FRONTEND_IMMEDIATES_SRCS)
 
-DYNO_IMMEDIATES_OBJS = \
-	$(DYNO_IMMEDIATES_SRCS:%.cpp=$(DYNO_IMMEDIATES_OBJDIR)/%.$(OBJ_SUFFIX))
+FRONTEND_IMMEDIATES_OBJS = \
+	$(FRONTEND_IMMEDIATES_SRCS:%.cpp=$(FRONTEND_IMMEDIATES_OBJDIR)/%.$(OBJ_SUFFIX))
