+++ frontend/lib/types/Makefile
@@ -17,11 +17,11 @@
 # limitations under the License.
 
 ifndef CHPL_MAKE_HOME
-export CHPL_MAKE_HOME=$(shell pwd)/../../../..
+export CHPL_MAKE_HOME=$(shell pwd)/../../..
 endif
 
-COMPILER_ROOT = ../../..
-COMPILER_SUBDIR = dyno/lib/types
+COMPILER_ROOT = ../../../compiler
+COMPILER_SUBDIR = frontend/lib/types
 
 #
 # standard header
@@ -30,7 +30,7 @@ include $(COMPILER_ROOT)/make/Makefile.compiler.head
 
 include Makefile.include
 
-TARGETS = $(DYNO_TYPES_OBJS)
+TARGETS = $(FRONTEND_TYPES_OBJS)
 
 include $(COMPILER_ROOT)/make/Makefile.compiler.subdirrules
