+++ tools/chpldoc/BUILD_VERSION
@@ -1,1 +1,1 @@
-../../../main/BUILD_VERSION
+../../compiler/main/BUILD_VERSION
