+++ tools/chpldoc/COPYRIGHT
@@ -1,1 +1,1 @@
-../../../main/COPYRIGHT
+../../compiler/main/COPYRIGHT
