+++ tools/chpldoc/LICENSE
@@ -1,1 +1,1 @@
-../../../main/LICENSE
+../../compiler/main/LICENSE
