+++ tools/chpldoc/arg.h
@@ -1,1 +1,1 @@
-../../../include/arg.h
+../../compiler/include/arg.h
