+++ tools/chpldoc/version_num.h
@@ -1,1 +1,1 @@
-../../../main/version_num.h
+../../compiler/main/version_num.h
