+++ util/buildRelease/gen_release
@@ -135,9 +135,9 @@ if (exists($ENV{"CHPL_GEN_RELEASE_NO_CLONE"})) {
        "util/setchplenv.sh",
        "util/start_test",
        "util/chpltags",
-       "compiler/dyno/tools/chpldoc/LICENSE",
-       "compiler/dyno/tools/chpldoc/COPYRIGHT",
-       "compiler/dyno/include/chpl/config/config.h.cmake",
+       "tools/chpldoc/LICENSE",
+       "tools/chpldoc/COPYRIGHT",
+       "frontend/include/chpl/config/config.h.cmake",
 );
 
 
@@ -224,8 +224,8 @@ print "[gen_release] Removing runtime directories that are not ready for release
 SystemOrDie("cd runtime/src/launch && rm -r dummy");
 SystemOrDie("cd runtime/src/launch && rm -r mpirun");
 
-print "[gen_release] Removing dyno test directory which is not intended for release...\n";
-SystemOrDie("cd compiler/dyno && rm -r test");
+print "[gen_release] Removing frontend test directory which is not intended for release...\n";
+SystemOrDie("cd frontend && rm -r test");
 
 print "[gen_release] Removing third-party directories that are not intended for release...\n";
 SystemOrDie("cd third-party && rm *.devel*");
